Add common method `prompt_for_password()` to create robust prompts
The `prompt_for_password()` function allows to asks for password input
twice and only returns `Some(Password)` if inputs match and are not
empty. If both inputs are empty, returns `None`.
The function takes a prompt string to provide to the user and an
optional prompt string for the repetition prompt.

Signed-off-by: David Runge <dave@sleepmap.de>